back_end/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
djoser==2.2.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1